requirements.txt
coverage==5.0.4
Django==3.1.3
django-extensions==3.0.9
djangorestframework==3.12.2
factory-boy==2.12.0
psycopg2==2.8.4
responses==0.12.0
requests==2.24.0
python-memcached==1.59
django-filter==2.4.0
gunicorn==20.0.4
django-heroku==0.3.1
django-heroku-memcacheify==1.0.0
pylibmc==1.5.2